+++ examples/vue/dynamic/src/components/GridVirtualizerDynamic.vue
@@ -9,7 +9,7 @@
       <template v-for="virtualRow in virtualRows" :key="virtualRow.key">
         <div
           :data-index="virtualRow.index"
-          :ref="measureElement"
+          ref="virtualItemEls"
           :style="{
             position: 'absolute',
             top: 0,
@@ -47,9 +47,9 @@
 </template>
 
 <script setup lang="ts">
-import { ref, computed, onMounted, type VNodeRef } from 'vue'
-import { useWindowVirtualizer, useVirtualizer } from '@tanstack/vue-virtual'
-import { generateData, generateColumns } from './utils'
+import { computed, onMounted, onUpdated, ref, shallowRef } from 'vue'
+import { useVirtualizer, useWindowVirtualizer } from '@tanstack/vue-virtual'
+import { generateColumns, generateData } from './utils'
 
 const columns = generateColumns(30)
 const data = generateData(columns)
@@ -103,13 +103,15 @@ const width = computed(() => {
     : [0, 0]
 })
 
-const measureElement = (el) => {
-  if (!el) {
-    return
-  }
+const virtualItemEls = shallowRef([])
 
-  rowVirtualizer.value.measureElement(el)
-
-  return undefined
+function measureAll() {
+  rowVirtualizer.value.measureElement(null)
+  virtualItemEls.value.forEach((el) => {
+    if (el) rowVirtualizer.value.measureElement(el)
+  })
 }
+
+onMounted(measureAll)
+onUpdated(measureAll)
 </script>